reduced-requirements 0.5.1.txt
# Core dependencies
python-dotenv==1.1.0
requests==2.32.3
pyttsx3==2.98
SpeechRecognition==3.14.2
schedule==1.2.2
vosk==0.3.44

# Machine Learning
numpy==2.2.5
pandas==2.2.3
scikit-learn==1.6.1
nltk==3.9.1

# Optional NLP/AI components
transformers==4.51.3
torch==2.7.0

# Web scraping and data processing
beautifulsoup4==4.13.4
wikipedia==1.4.0

# Utilities
tqdm==4.67.1
rich==14.0.0

# Date handling
python-dateutil==2.9.0.post0
pytz==2025.2

# Configuration
PyYAML==6.0.2

# Mac specific (only for macOS)
pyobjc==11.0